Transform debugging | Outlining objects
-Fixed some issues with the child gizmos not appearing in their respective position when the parent was transformed.
NOTE: now there is another bug where the position displayed is always relative to the parent and not the globally and the position is not updated respect to the world when parent is moved.

-Added some outlining for game object selection.
BUGS: since the position of the child is not updated in the world, the outlining has an offset.

-Starting scene is the Street1.fbx.

-Tried to add a logo in window.